--- src/main/java/com/hamter/model/Allcodes.java
@@ -1,82 +1,0 @@
-package com.hamter.model;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "allcodes")
-public class AllCode {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String key;
-
-    private String type;
-
-    private String valueEn;
-
-    private String valueVi;
-
-    @Column(name = "createdAt", nullable = false)
-    private java.util.Date createdAt;
-
-    @Column(name = "updatedAt", nullable = false)
-    private java.util.Date updatedAt;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getKey() {
-        return key;
-    }
-
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getValueEn() {
-        return valueEn;
-    }
-
-    public void setValueEn(String valueEn) {
-        this.valueEn = valueEn;
-    }
-
-    public String getValueVi() {
-        return valueVi;
-    }
-
-    public void setValueVi(String valueVi) {
-        this.valueVi = valueVi;
-    }
-
-    public java.util.Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(java.util.Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public java.util.Date getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(java.util.Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
