+++ src/main/java/com/hamter/model/Booking.java
@@ -1,0 +1,37 @@
+package com.hamter.model;
+
+import javax.persistence.*;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@Entity
+@Table(name = "bookings")
+public class Booking {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String statusId;
+
+    private Integer doctorId;
+
+    private String patientId;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date date;
+
+    private String timeType;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "createdAt", nullable = false)
+    private Date createdAt;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updatedAt", nullable = false)
+    private Date updatedAt;
+
+}
