+++ src/main/java/com/hamter/model/Clinic.java
@@ -1,0 +1,35 @@
+package com.hamter.model;
+
+import javax.persistence.*;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@Entity
+@Table(name = "clinics")
+public class Clinic {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    private String address;
+
+    @Lob
+    private String description;
+
+    private String image;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "createdAt", nullable = false)
+    private Date createdAt;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updatedAt", nullable = false)
+    private Date updatedAt;
+ 
+}
