+++ src/main/java/com/hamter/model/Doctor_Clinic_Specialty.java
@@ -1,0 +1,5 @@
+package com.hamter.model;
+
+public class Doctor_Clinic_Specialty {
+
+}
