+++ src/main/java/com/hamter/model/History.java
@@ -1,0 +1,36 @@
+package com.hamter.model;
+
+import javax.persistence.*;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@Entity
+@Table(name = "histories")
+public class History {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Integer patientId;
+
+    private Integer doctorId;
+
+    @Lob
+    private String description;
+
+    @Lob
+    private String files;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "createdAt", nullable = false)
+    private Date createdAt;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updatedAt", nullable = false)
+    private Date updatedAt;
+
+}
