+++ src/main/java/com/hamter/model/Schedule.java
@@ -1,0 +1,37 @@
+package com.hamter.model;
+
+import javax.persistence.*;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@Entity
+@Table(name = "Schedules")
+public class Schedule {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Integer currentNumber;
+
+    private Integer maxNumber;
+
+    @Temporal(TemporalType.DATE)
+    private Date date;
+
+    private String timeType;
+
+    private Integer doctorId;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "createdAt", nullable = false)
+    private Date createdAt;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updatedAt", nullable = false)
+    private Date updatedAt;
+
+}
