+++ src/main/java/com/hamter/model/Specialty.java
@@ -1,0 +1,34 @@
+package com.hamter.model;
+
+import javax.persistence.*;
+
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@Entity
+@Table(name = "specialties")
+public class Specialty {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @Lob
+    private String description;
+
+    @Lob
+    private String image;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "createdAt", nullable = false)
+    private Date createdAt;
+
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "updatedAt", nullable = false)
+    private Date updatedAt;
+
+}
