+++ src/main/resources/application.properties
@@ -1,1 +1,13 @@
 spring.application.name=Clinic_Management
+
+spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=ClinicManagement;encrypt=true;trustServerCertificate=true;
+spring.datasource.username=sa
+spring.datasource.password=123456
+spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
